patching the router.ex with users resources
Not sure about this one.  The current guide says to put the `resources
"/users"` inside the api pipe.  But the routing guide seems to
suggest that `resources "/users" ...` belongs in the `scope /api` block
instead.  So I moved it there and mentioned uncommenting that block.

Also, use the `except` option (something new here) because we don't
need to get forms generated for a JSON API.  Noted as such in text